Latest working release of grayscale routine (waiting for Kevin one).
git-svn-id: http://svn.tilp.info/repos/tiemu@829 4f589446-7ac1-0310-82c7-a055b08196ef

+++ tiemu/trunk/src/core/ti68k_def.h
@@ -127,7 +127,7 @@ typedef struct
 	int		log_w;			// LCD logical width
 	int		log_h;			// LCD logical height
 	int		on_off;
-	int		lcd_tick;		// used by grayscales
+	unsigned long	lcd_tick;		// used by grayscales
 
     // linkport.c
 	int		lc_speedy;		// not used any longer (free of use)
